+++ lib/bytes.dart
@@ -1,3 +1,5 @@
+import 'package:meta/meta.dart';
+
 import 'builder_helper.dart' as helper;
 import 'parser_builder.dart';
 
@@ -8,6 +10,7 @@ part 'src/bytes/tag.dart';
 part 'src/bytes/tag_no_case.dart';
 part 'src/bytes/tag_of.dart';
 part 'src/bytes/tags.dart';
+part 'src/bytes/tags_values.dart';
 part 'src/bytes/take_until.dart';
 part 'src/bytes/take_until1.dart';
 part 'src/bytes/take_while.dart';